Modified Task 1.1

+++ src/lesson1/task1/Simple.kt
@@ -71,7 +71,7 @@ fun lengthInMeters(sagenes: Int, arshins: Int, vershoks: Int): Double =
  * Вывести значение того же угла в радианах (например, 0.63256).
  */
 fun angleInRadian(grad: Int, min: Int, sec: Int): Double =
-        (grad + min.toDouble() / 60 + sec.toDouble() / 3600) * PI / 180
+        (grad + min / 60.0 + sec / 3600.0) * PI / 180
 
 /**
  * Тривиальная
